Refactor and enhance system architecture and rendering pipeline
- Added OnFixedUpdate method to System interface for fixed timestep updates.
- Updated SystemManager to iterate through systems using an ordered list.
- Improved serialization error logging in SerializationManager.
- Refactored CollisionSystem to handle collision resolution and detection more effectively, including support for multiple shapes. WIP
- Enhanced PhysicsSystem to initialize Box2D world and manage physics updates.
- Updated RenderSystem to improve texture rendering and culling logic based on camera position.
- Added CameraSystem functionality to calculate screen positions from world coordinates.
- Refactored TransformSystem to streamline entity transformations and hierarchy management.
- Updated TODO list to reflect current development priorities and tasks.

+++ HotBeanEngine/src/components/rendering/texture.hpp
@@ -16,7 +16,7 @@
 namespace HBE::Components {
     struct Texture : public Component {
         SDL_Texture* m_texture = nullptr; ///< A pointer to the SDL texture object. Can be null if the texture has not been loaded.
-        glm::ivec2 m_size = {0, 0}; ///< Size of the texture in pixels.
+        glm::vec2 m_size = {0, 0}; ///< Size of the texture in pixels.
         bool m_dirty = true;
 
         Texture() = default;
